+++ xenon-common/src/main/java/com/vmware/xenon/common/NodeSelectorState.java
@@ -13,10 +13,50 @@
 
 package com.vmware.xenon.common;
 
+
+import java.util.EnumSet;
+
+import com.vmware.xenon.services.common.NodeGroupService.NodeGroupState;
+import com.vmware.xenon.services.common.NodeGroupUtils;
+
+
 public class NodeSelectorState extends ServiceDocument {
+
+    public enum Status {
+        UNAVAILABLE,
+        PAUSED,
+        AVAILABLE
+    }
+
+    public static final EnumSet<Status> UNAVAILABLE = EnumSet.of(Status.UNAVAILABLE);
+    public static final EnumSet<Status> PAUSED = EnumSet.of(Status.PAUSED, Status.AVAILABLE);
+    public static final EnumSet<Status> AVAILABLE = EnumSet.of(Status.AVAILABLE);
+    public static final EnumSet<Status> PAUSED_UNAVAILABLE = EnumSet.of(Status.PAUSED, Status.UNAVAILABLE);
+
+    /**
+     * Calculates the status of the Node Selector by inspecting the NodeGroup and local
+     * 'pause'/'resume' status.
+     */
+    public static Status calculateStatus(ServiceHost host, NodeGroupState localState) {
+        boolean isAvailable = NodeGroupUtils.isNodeGroupAvailable(host, localState);
+        return (isAvailable ? Status.AVAILABLE : Status.UNAVAILABLE);
+    }
+
+    public static void updateStatus(NodeSelectorState ns, ServiceHost host, NodeGroupState localState) {
+        ns.nodeSelectorStatus = calculateStatus(host, localState);
+    }
+
+    public static boolean isAvailable(NodeSelectorState ns) {
+        return AVAILABLE.contains(ns.nodeSelectorStatus);
+    }
+
+    public static boolean isAvailable(ServiceHost host, NodeGroupState localstate) {
+        return AVAILABLE.contains(calculateStatus(host, localstate));
+    }
+
     public String nodeGroupLink;
     public Long replicationFactor;
     public int membershipQuorum;
     public long membershipUpdateTimeMicros;
-    public boolean isNodeGroupAvailable;
+    public Status nodeSelectorStatus;
 }